Split statement-specific parts of `semantics_handle.cpp` into a couple of separate files.
